Move comments and issues logic into separate packages
